fix relative paths with double separators
Computing the relative path for "/e/b//c/d/"
if the cwd is "/e/b/" would not result in a
correct relative path (extra leading slash).

This fixes that by cleaning the path first.
This was partly done when normalizing the path,
but now that is no longer needed, as it's done
before the home directory is replaced in the
relative path step of the normalization.

Extra tests were added to include cases with
double separators, and make them OS-independent.