feat(profile): add wrapper around LuaJIT profiler
This is a thin wrapper around LuaJIT's `jit.p` with the right "magic"
options set. Only configuration right now is `flame=true` to output
flamegraph-compatible logs; see https://github.com/jonhoo/inferno
to be used with
```
inferno-flamegraph profile.log > flame.svg
```
The SVG file can then be viewed in the browser.

This commit includes a vendored version of the latest upstream `jit.p`
that fixes a bug with flushing the output file; once
https://github.com/LuaJIT/LuaJIT/commit/95140c50010c0557af66dac944403a1a65dd312c
is pulled in, this file can be removed.