fix: ensure `Border:change_title` doesn't error
I believe this was a regression introduced in
https://github.com/nvim-lua/plenary.nvim/commit/8c6cc07a68b65eb707be44598f0084647d495978,
but currently when you call `Border:change_title` you'll see:

```
E5108: Error executing lua .../packer/start/plenary.nvim/lua/plenary/window/border.lua:57: attempt to in
dex local 'thickness' (a nil value)
```
This is because after `Border:__align_calc_config` is called, the new
`border_win_options` were never actually updated in the metatable. This
ensure that after the table is merged that it's correctly updated so
that when `Border:change_title` is called, it still has the correct
reference to these.