fix: filetype detection freeze with binary files
Filetype detection used to read the first line for shebang and the last
line for modeline detection. This was really slow with large binary
files that don't contain newlines.

Add Path:readbyterange(offset, length) to allow reading a specific range
and reimplement detect_from_modeline and detect_from_shebang to work
with the first 256 and last 256 bytes of a file.

Fixes #307 and closes #322.