fix: path normalization works with esoteric usernames (#290)
* fix: path normalization works with esoteric usernames

Usernames with dashes at the end are perfectly valid unix user names,
but they're seen as a regular expression pattern in Lua. Switch to using
string.find instead of string.gsub to avoid the pattern matching issue.

* Fix syntax error

Using end here is no good, even though it worked in the luajit repl. 😆

* Fix typo in spec name

* Clean the path we send to _normalize_path

Also fix the specs to match the new behavior.