Handle errors in callback passed to `context_manager.with`
This more closely emulates the behavior of the Python `with` statement,
which according to comments this is modeled off of.

With this change the context will always be closed, even if an error
occurs. This means that, for example, you can open a file without having
to worry about an error leaking the file handle. Or if you created a
context for creating temporary files and deleting them when done, you
wouldn't need to worry about failing to delete the temp file.

This change also re-enables disabled `context_manager` tests.

Fixes #340