feat: add `memoize.lua`.
Credits goes to: https://github.com/kikito/memoize.lua

This add a `memoize` function that lets users cache return values from a
function passed as an argument.

It is pretty much a copy paste of the that `kikito` already wrote, with some
changes to follow the `stylua` formatting.

This technically could go to `functional.lua`, but it would be probably better
to have it in its own file.