Fix PlenaryBustedDirectory command
expand() should only be called for the first argument, not for the
second one.
The second argument is a string representation of a Lua option table
which did work in case of a single entry, where only the braces were
removed.
For multiple entries Lua was not able to parse the expanded version as
option table anymore, though.