Adjusted path sep logic
Added support for Windows-style paths with forward slash (i.e. when `vim.fn.exists('+shellslash')` and `vim.opt.shellslash._value` are both `true`). Previously the code assumed that sep would be '/' iff the OS was not Windows. I have reworked the code to avoid this assumption.

I also added an additional call to the built-in path normalizer at the end of the Plenary normalization function, just in case.