fix(log): Use `stdpath('log')` (#607)
Neovim has had a standard path for log files since `v0.7.0-322-g78a1e6bc0`, so this should be the preferred location. Logs aren't cache data, so they shouldn't be stored with cache data.

Given that plenary is used in a lot of plugins, this would be a tiny but free improvement in correctness for those plugins.