neovim 0.11 introduced winborder: neovim/neovim#31074
This means that the user may override the global window border for all
windows for which the border wasn't explicitly set.
For the scrollbar and the scrollbar thumb, we don't want any border.

In the past we achieved that by now specifying a border, but with
winborder, the user may set the border and then the scrollbar looks
broken. So now we explicitly say that we want no border on these two
windows, fixing the issue.