Verbose changes in 'path.lua' module to be sure that the path will be in windows-style but due to possible skips of the Path module we can get posix-style path in 'job.lua' module as well.
In my case I get stable Windows-style paths in 'harpoon' menu (no mixed slashes).
The changes allow me to use git-related functions from Telescope but it
also needs to be changed a bit cause it invokes 'vim.loop.spawn()' directly
that leads to posix-style paths in the final command which WinAPI rejects.

The changes are experimental cause integration with msys2 is fragile.